- Added the shader background to all the components - Removed unnecessary asset files - Fixed broken imports in OurTeam.jsx and index.
