+++ src/App.css
@@ -1,12 +1,88 @@
-body{
-    background: 
-      radial-gradient(circle at bottom right, rgba(128, 0, 128, 0.5), transparent 40%),
-      linear-gradient(to bottom left, #0d1f2d, #0d1f2d,#050551);
+/* Global body transparency for shader background */
+body {
+  background: transparent !important;
 }
 
-.bg-grid {
-  background: 
-    linear-gradient(to right, rgba(255, 255, 255, 0.1) 1px, transparent 1px),
-    linear-gradient(to bottom, rgba(255, 255, 255, 0.1) 1px, transparent 1px);
-  background-size: 100px 100px; /* Adjust grid spacing */
+/* Remove blocking backgrounds - previous rules */
+.bg-\[#0a192f\],
+.bg-\[#061529\],
+.bg-\[#0d1f2d\] {
+  background: transparent !important;
+}
+
+/* Events Section - Remove all backgrounds */
+.events-section,
+.live-events-container,
+.live-events-panel-desktop {
+  background: transparent !important;
+  background-color: transparent !important;
+}
+
+/* Gallery Section - Remove all backgrounds */
+.gallery-background,
+.gallery-container,
+.gallery-mobile-content {
+  background: transparent !important;
+  background-color: transparent !important;
+  background-image: none !important;
+}
+
+/* App container override */
+.app-container {
+  background: transparent !important;
+  background-color: transparent !important;
+}
+
+/* Override specific gradient and color backgrounds */
+*[style*="background: linear-gradient"],
+*[style*="background-color: #0b0f19"],
+*[style*="background-color: #000c18"],
+*[style*="background-color: #0d1a2e"],
+*[style*="background-color: #1a2c47"] {
+  background: transparent !important;
+}
+
+/* Ensure all sections allow shader to show through */
+section,
+main,
+div[class*="section"],
+div[class*="container"] {
+  background: transparent !important;
+}
+
+/* Keep card/element functionality but make transparent */
+.event-card {
+  background-color: rgba(17, 24, 39, 0.1) !important;
+}
+
+.event-card-image {
+  background-color: rgba(255, 255, 255, 0.05) !important;
+}
+
+.gallery-image {
+  background-color: rgba(160, 160, 160, 0.1) !important;
+}
+
+/* Keep navigation functionality */
+.nav-arrow {
+  background-color: rgba(255, 255, 255, 0.1) !important;
+}
+
+.nav-arrow:hover {
+  background-color: rgba(255, 255, 255, 0.2) !important;
+}
+
+/* Keep button backgrounds functional but transparent */
+.details-button {
+  background-color: rgba(229, 231, 235, 0.8) !important;
+}
+
+.tag-button {
+  background-color: rgba(26, 44, 71, 0.3) !important;
+}
+
+/* Sponsors section transparency */
+.sponsi-container,
+.Sponser {
+  background: transparent !important;
 }
