+++ src/components/Hero.jsx
@@ -1,153 +1,7 @@
 "use client"
 import { useEffect, useRef, useState } from "react"
-import * as THREE from "three"
 import NSCCEvectorImg from "../assets/NSCC EVECTOR.png"
 import nscchero from "../assets/hero.png"
-
-const ShaderBackground = () => {
-  const mountRef = useRef(null)
-  const animationRef = useRef(null)
-
-  useEffect(() => {
-    if (!mountRef.current) return
-
-    const scene = new THREE.Scene()
-    const camera = new THREE.PerspectiveCamera(75, window.innerWidth / window.innerHeight, 0.1, 1000)
-    const renderer = new THREE.WebGLRenderer({
-      alpha: true,
-      antialias: true,
-      preserveDrawingBuffer: true,
-      premultipliedAlpha: false,
-    })
-
-    renderer.setSize(window.innerWidth, window.innerHeight)
-    renderer.setPixelRatio(Math.min(window.devicePixelRatio, 2))
-    mountRef.current.appendChild(renderer.domElement)
-
-    const material = new THREE.ShaderMaterial({
-      fragmentShader: originalFragmentShader,
-      vertexShader: originalVertexShader,
-      uniforms: {
-        u_color: { value: [0.3137254901960784, 0, 1] },
-        u_background: { value: [0.039, 0.098, 0.184, 1] },
-        u_speed: { value: 0.876 },
-        u_detail: { value: 0.074 },
-        u_time: { value: 0 },
-        u_mouse: { value: [0, 0] },
-        u_resolution: { value: [window.innerWidth, window.innerHeight] },
-      },
-      transparent: true,
-    })
-
-    const geometry = new THREE.PlaneGeometry(1, 1)
-    const mesh = new THREE.Mesh(geometry, material)
-    scene.add(mesh)
-    mesh.position.z = -1
-    camera.position.z = 0
-
-    const resizePlane = () => {
-      const distance = Math.abs(mesh.position.z - camera.position.z)
-      const fovInRadians = (camera.fov * Math.PI) / 180
-      const height = 2 * Math.tan(fovInRadians / 2) * distance
-      const width = height * camera.aspect
-      mesh.scale.set(width, height, 1)
-    }
-
-    resizePlane()
-
-    const mouse = { x: 0, y: 0 }
-
-    const handleMouseMove = (event) => {
-      mouse.x = (event.clientX / window.innerWidth) * 2 - 1
-      mouse.y = -(event.clientY / window.innerHeight) * 2 + 1
-    }
-
-    window.addEventListener("mousemove", handleMouseMove)
-
-    const clock = new THREE.Clock()
-    const animate = () => {
-      const elapsedTime = clock.getElapsedTime()
-      material.uniforms.u_time.value = elapsedTime
-      material.uniforms.u_mouse.value = [mouse.x / 2 + 0.5, mouse.y / 2 + 0.5]
-      material.uniforms.u_resolution.value = [window.innerWidth, window.innerHeight]
-      renderer.render(scene, camera)
-      animationRef.current = requestAnimationFrame(animate)
-    }
-
-    animate()
-
-    const handleResize = () => {
-      camera.aspect = window.innerWidth / window.innerHeight
-      camera.updateProjectionMatrix()
-      renderer.setSize(window.innerWidth, window.innerHeight)
-      material.uniforms.u_resolution.value = [window.innerWidth, window.innerHeight]
-      resizePlane()
-    }
-
-    window.addEventListener("resize", handleResize)
-
-    return () => {
-      window.removeEventListener("mousemove", handleMouseMove)
-      window.removeEventListener("resize", handleResize)
-      if (animationRef.current) cancelAnimationFrame(animationRef.current)
-      if (mountRef.current && renderer.domElement) mountRef.current.removeChild(renderer.domElement)
-      geometry.dispose()
-      material.dispose()
-      renderer.dispose()
-    }
-  }, [])
-
-  return <div ref={mountRef} className="absolute inset-0 z-0 opacity-80" />
-}
-
-// Your original fragment shader
-const originalFragmentShader = `
-uniform vec2 u_resolution;
-uniform float u_time;
-uniform vec3 u_color;
-uniform vec4 u_background;
-uniform float u_speed;
-uniform float u_detail;
-
-mat2 m(float a) {
-  float c=cos(a), s=sin(a);
-  return mat2(c,-s,s,c);
-}
-
-float map(vec3 p) {
-  float t = u_time * u_speed;
-  p.xz *= m(t * 0.4);
-  p.xy*= m(t * 0.1);
-  vec3 q = p * 2.0 + t;
-  return length(p+vec3(sin((t*u_speed) * 0.1))) * log(length(p) + 0.9) + cos(q.x + sin(q.z + cos(q.y))) * 0.5 - 1.0;
-}
-
-void main() {
-  vec2 a = gl_FragCoord.xy / u_resolution.x - vec2(0.5, 0.5);
-  vec3 cl = vec3(0.0);
-  float d = 2.5;
-  for (float i = 0.; i <= (1. + 20. * u_detail); i++) {
-    vec3 p = vec3(0, 0, 4.0) + normalize(vec3(a, -1.0)) * d;
-    float rz = map(p);
-    float f = clamp((rz - map(p + 0.1)) * 0.5, -0.1, 1.0);
-    vec3 l = vec3(0.1, 0.3, 0.4) + vec3(5.0, 2.5, 3.0) * f;
-    cl = cl * l + smoothstep(2.5, 0.0, rz) * 0.6 * l;
-    d += min(rz, 1.0);
-  }
-  vec4 color = vec4(min(u_color, cl),1.0);
-  color.r = max(u_background.r,color.r);
-  color.g = max(u_background.g,color.g);
-  color.b = max(u_background.b,color.b);
-  gl_FragColor = color;
-}
-`
-
-// Your original vertex shader
-const originalVertexShader = `
-void main() {
-  gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
-}
-`
 
 export default function Hero() {
   const [isMenuOpen, setIsMenuOpen] = useState(false)
@@ -185,19 +39,16 @@ export default function Hero() {
   }, [isMenuOpen])
 
   return (
-    <div ref={heroRef} className="relative min-h-screen bg-[#0a192f] overflow-x-hidden">
-      <ShaderBackground />
-
-      {/* Grid Lines - Limited to hero section only */}
+    <div ref={heroRef} className="relative min-h-screen overflow-x-hidden">
+      {/* Grid Lines - 4x3 grid matching screenshot */}
       <div className="absolute inset-0 z-0 hidden lg:block">
-        {/* Vertical Lines */}
+        {/* Vertical Lines - 4 columns (3 dividing lines) */}
         <div className="absolute top-0 left-1/4 w-[1px] h-full bg-gradient-to-b from-transparent via-white/20 to-transparent"></div>
         <div className="absolute top-0 left-1/2 w-[1px] h-full bg-gradient-to-b from-transparent via-white/20 to-transparent"></div>
         <div className="absolute top-0 left-3/4 w-[1px] h-full bg-gradient-to-b from-transparent via-white/20 to-transparent"></div>
-        {/* Horizontal Lines */}
-        <div className="absolute top-1/4 left-0 w-full h-[1px] bg-gradient-to-r from-transparent via-white/20 to-transparent"></div>
-        <div className="absolute top-1/2 left-0 w-full h-[1px] bg-gradient-to-r from-transparent via-white/20 to-transparent"></div>
-        <div className="absolute top-3/4 left-0 w-full h-[1px] bg-gradient-to-r from-transparent via-white/20 to-transparent"></div>
+        {/* Horizontal Lines - 3 rows (2 dividing lines) */}
+        <div className="absolute top-1/3 left-0 w-full h-[1px] bg-gradient-to-r from-transparent via-white/20 to-transparent"></div>
+        <div className="absolute top-2/3 left-0 w-full h-[1px] bg-gradient-to-r from-transparent via-white/20 to-transparent"></div>
       </div>
 
       {/* Navigation */}
@@ -253,7 +104,7 @@ export default function Hero() {
       >
         {/* Backdrop */}
         <div
-          className={`absolute inset-0 bg-[#0a192f]/100 transition-opacity duration-500 ${isMenuOpen ? "opacity-80" : "opacity-0"}`}
+          className={`absolute inset-0 bg-[#0a192f]/95 transition-opacity duration-500 ${isMenuOpen ? "opacity-80" : "opacity-0"}`}
         ></div>
 
         {/* Menu Content */}
@@ -335,9 +186,9 @@ export default function Hero() {
         {/* Mobile/Tablet version of Lorem Ipsum */}
         <div className="xl:hidden w-full max-w-md mx-auto mt-4 px-4">
           <p className="text-sm md:text-base text-gray-300 leading-relaxed font-helvetica-neue text-center">
-            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et
-            dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex
-            ea commodo consequat.
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt 
+            ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco 
+            laboris nisi ut aliquip ex ea commodo consequat.
           </p>
         </div>
       </section>
