+++ src/components/events/Events.css
@@ -4,7 +4,7 @@
 }
 
 .app-container {
-  background-color: #000c18;
+  background-color: transparent;
 }
 
 /* Live Events Section */
@@ -16,7 +16,7 @@
   align-items: center;
   justify-content: center;
   text-align: center;
-  background: linear-gradient(to bottom, #000c18, #001a36);
+  background: transparent;
   padding: 0 2rem; /* Add some padding for content */
 }
 
@@ -39,7 +39,7 @@
 .events-section {
   width: 100%;
   color: white;
-  background-color: #0b0f19; /* Default background for desktop */
+  background-color: transparent; /* Default background for desktop */
 }
 
 .event-carousel-wrapper {
@@ -95,7 +95,7 @@
     height: 500px;
     margin: 0 2rem;
     flex-shrink: 0;
-    background-color: #0d1a2e;
+    background-color: rgba(13, 26, 46, 0.1);
     border: 1px solid #253a5e;
     border-radius: 1.5rem;
     display: flex;
@@ -107,7 +107,7 @@
 
   .event-card-image {
     height: 45%;
-    background-color: #1a2c47;
+    background-color: rgba(26, 44, 71, 0.1);
     border-radius: 1rem;
   }
 
@@ -158,7 +158,7 @@
   }
 
   .event-card .tag-button:hover {
-    background-color: #1a2c47;
+    background-color: rgba(26, 44, 71, 0.1);
     border-color: #a0aec0;
   }
 
@@ -171,7 +171,7 @@
 /* Mobile specific styles */
 @media (max-width: 767px) {
   .events-section {
-    background: linear-gradient(160deg, #0a192f 0%, #1d3b5a 50%, #2c5a8e 100%);
+    background: transparent;
     min-height: 100vh;
     padding: 4rem 1rem;
     display: flex;
