+++ src/components/gallery/Gallery.css
@@ -16,7 +16,7 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  background-color: #0b0f19;
+  background-color: transparent;
   background-image: linear-gradient(rgba(255, 255, 255, 0.03) 1px, transparent 1px),
     linear-gradient(90deg, rgba(255, 255, 255, 0.03) 1px, transparent 1px);
   background-size: 50px 50px;
@@ -104,7 +104,7 @@
     border-radius: 0.5rem;
     overflow: hidden;
     box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-    background-color: #a0a0a0;
+    background-color: rgba(160, 160, 160, 0.1);
   }
 
   .gallery-mobile-image {
